MM-22417: Clarify behavior of /channels/{channel_id}/posts endpoint
It is known that the posts are actually returned from their update time
and not create time, and there can be gaps in the post order.

Although the gaps are not expected, but it is not known what will break
at this stage, so we just fix the documentation to clarify the behavior.