Clarify how to change system configuration
It took me a while to realize that the whole configuration should
be specified when doing a PUT (not just the fields that one wants
to change). Here is a small paragraph explaining that, hoping to
help others.

It looks like secret parameters (passwords, keys?) are automatically
replaced with strings of asterisks (`**********...`) and that these
get ignored when we PUT a new configuration, but I'm not sure.
If someone knows for sure, it would be fantastic to clarify as well.

Thanks! ♥